src/backend/requirements.txt
asgiref==3.8.1
attrs==24.3.0
autobahn==24.4.2
Automat==24.8.1
black==24.10.0
certifi==2024.12.14
cffi==1.17.1
channels==4.2.0
channels_redis==4.2.1
charset-normalizer==3.4.0
click==8.1.8
cloudinary==1.41.0
constantly==23.10.4
cryptography==44.0.0
daphne==4.1.2
dj-database-url==2.3.0
Django==5.0
django-cors-headers==4.3.1
django-heroku==0.3.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
gunicorn==23.0.0
hvac==2.3.0
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
msgpack==1.1.0
mypy-extensions==1.0.0
packaging==24.2
pathspec==0.12.1
platformdirs==4.3.6
psycopg2==2.9.10
psycopg2-binary==2.9.9
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
PyJWT==2.10.1
pyOpenSSL==25.0.0
pyotp==2.9.0
python-dotenv==1.0.0
pytz==2024.2
qrcode==8.0
redis==5.2.1
requests==2.32.3
service-identity==24.2.0
six==1.17.0
sqlparse==0.5.3
Twisted==24.11.0
txaio==23.1.1
typing_extensions==4.12.2
urllib3==2.2.3
whitenoise==6.6.0
zope.interface==7.2
django-background-tasks
gunicorn
django-heroku
channels-redis
dj-database-url
